Move requestIdleCallback shim to its own file
